requirements.txt
rich
requests
dnspython
